backend/requirements.txt
# Web Framework
fastapi>=0.109.2
uvicorn[standard]>=0.27.1

# Database
sqlalchemy[asyncio]>=2.0.27
psycopg[binary]>=3.1.18
asyncpg>=0.29.0
alembic>=1.13.1

# Utils
python-dotenv>=1.0.1
pydantic[email]>=2.6.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
